Added background music to FMOD
Added an additional event in the FMOD project harpoonsfx.fspro with
parameters for the music. The “Level” parameter defaults to 0.

Setting it to 1 will transition from the title theme to the calm music.

Setting it to 2 will transition from the calm music to storm music.

Setting it to 3 will transition from the storm music to the calm music.