Music transition to title theme
Added a transition back to the title screen music. Setting the “Level”
parameter to 0 at any point during the game will transition to the
start of the title theme loop.

Also, simplified the parameters: setting it to 1 always transitions to
the calm music, setting it to 2 always transitions to the storm music.